👌 IMPROVE: Rename option
